fix(config): Tool-compatible TypeScript configuration.
tsserver does not allow specifying a tsconfig path. This has an impact on some tools such as editors -- test files think that project-relative imports are invalid.

To address this, we use the tsconfig.json as the development/testing configuration. For building we have a separate configuration. react-native-typescript-transformer makes it easy to change and enforce.